+++ src/components/Modal/ModalDelete.jsx
@@ -3,6 +3,7 @@ import {
   ModalNegativeButton,
   ModalShareFolderTitle,
   ModalShareTitleContainer,
+  ModalTitle,
   ModalWrapper,
 } from "./Modal.styled";
 
@@ -10,7 +11,7 @@ export default function ModalDelete({ folderName = "폴더명" }) {
   return (
     <ModalWrapper>
       <ModalShareTitleContainer>
-        폴더 삭제
+        <ModalTitle>폴더 삭제</ModalTitle>
         <ModalShareFolderTitle>{folderName}</ModalShareFolderTitle>
       </ModalShareTitleContainer>
       <ModalNegativeButton>삭제하기</ModalNegativeButton>
